test: Move shuntingYard tests to new file
